xhci控制器初始化（还存在bug，无法reset端口）

+++ kernel/common/time.h
@@ -25,7 +25,7 @@ struct tm
 struct timespec
 {
     long int tv_sec;    // 秒
-    long int tv_nsec;   // 纳秒
+    long long tv_nsec;   // 纳秒
 };
 
 /**
